Code cleanup + shutdown pin
Added a function to set the shutdown pin low when turning off the device. This cuts the power to the pi after 3 seconds, while the pi shuts down. Otherwise you have to hold down the middle button for 3 seconds to cut power before you can turn it back on.

Code cleanup in various files.

Fixed charging screen battery level.

Added haptic feedback for long press on buttons.

+++ software/remote/sources/bq27441.h
@@ -89,7 +89,7 @@ public:
     Q_INVOKABLE int16_t  getAverageCurrent();
     int16_t  getStandbyCurrent();
     int16_t  getMaxLoadCurrent();
-    Q_INVOKABLE int16_t  getAveragePower();
+    Q_INVOKABLE int getAveragePower();
     Q_INVOKABLE int getStateOfCharge();
     uint16_t getInternalTemperature();            // Result in 0.1 Kelvins
     Q_INVOKABLE int16_t  getInternalTemperatureC();           // Result in 0.1 Celsius
